Refactor code and reformat style

+++ src/util/check_if_data_is_waiting.c
@@ -1,5 +1,5 @@
 int32_t check_if_data_is_waiting (const fd_set rfds) {
-    for (int32_t k = 0; k <= get_num_of_players(); ++k) {
+    for (size_t k = 0; k <= get_num_of_players(); ++k) {
         if (FD_ISSET(get_player_socket(k), &rfds)) {
             process_data(get_player_socket(k));
         }
